requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
exceptiongroup==1.1.1
Faker==18.10.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Pillow==9.5.0
pyaes==1.6.1
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
